refactor: jwt 파일구조 정리 및 리프래시 토큰 제외하여 깃허브 로그인 재구현
