--- src/main/java/com/leets/commitatobe/domain/login/presentation/dto/JwtResponse.java
@@ -1,8 +1,0 @@
-package com.leets.commitatobe.domain.login.presentation.dto;
-
-
-public record JwtResponse(
-    String grantType,
-    String accessToken,
-    String refreshToken
-) {}
