+++ src/main/java/com/leets/commitatobe/global/config/SecurityConfig.java
@@ -1,8 +1,8 @@
 package com.leets.commitatobe.global.config;
 
-import com.leets.commitatobe.global.filter.JwtAuthenticationFilter;
-import com.leets.commitatobe.global.utils.JwtProvider;
-import lombok.RequiredArgsConstructor;
+import java.util.Arrays;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -18,60 +18,63 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import java.util.Arrays;
-import java.util.List;
+import com.leets.commitatobe.global.jwt.filter.JwtAuthenticationFilter;
+import com.leets.commitatobe.global.jwt.provider.JwtProvider;
+
+import lombok.RequiredArgsConstructor;
 
 @Configuration
 @EnableWebSecurity
 @RequiredArgsConstructor
 public class SecurityConfig {
-    private final JwtProvider jwtProvider;
+	private final JwtProvider jwtProvider;
 
-    @Value("${management.endpoints.web.base-path}")
-    private String actuatorEndPoint;
+	@Value("${management.endpoints.web.base-path}")
+	private String actuatorEndPoint;
 
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
+	@Bean
+	public PasswordEncoder passwordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
 
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .httpBasic(AbstractHttpConfigurer::disable)
-                .cors((cors) -> cors
-                        .configurationSource(corsConfigurationSource())
-                )
-                .csrf(AbstractHttpConfigurer::disable)
-                .formLogin((formLogin) -> formLogin
-                        .loginPage("/login/github"))
-                .authorizeHttpRequests((authorize) ->
-                        authorize
-                                .requestMatchers("/v3/api-docs/**", "/swagger-ui/**", "/commit/**",
-                                        "/login/**", "/auth/**", "/h2-console/**", "/error/**", actuatorEndPoint + "/**").permitAll()
-                                .anyRequest().authenticated()
-                )
-                .headers(headers -> headers
-                        .frameOptions(HeadersConfigurer.FrameOptionsConfig::disable) // H2 콘솔 프레임 옵션 설정
-                )
-                .addFilterBefore(new JwtAuthenticationFilter(jwtProvider), UsernamePasswordAuthenticationFilter.class);
-        return http.build();
-    }
+	@Bean
+	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+		http
+			.httpBasic(AbstractHttpConfigurer::disable)
+			.cors((cors) -> cors
+				.configurationSource(corsConfigurationSource())
+			)
+			.csrf(AbstractHttpConfigurer::disable)
+			.formLogin((formLogin) -> formLogin
+				.loginPage("/login/github"))
+			.authorizeHttpRequests((authorize) ->
+				authorize
+					.requestMatchers("/v3/api-docs/**", "/swagger-ui/**", "/commit/**",
+						"/login/**", "/auth/**", "/h2-console/**", "/error/**", actuatorEndPoint + "/**").permitAll()
+					.anyRequest().authenticated()
+			)
+			.headers(headers -> headers
+				.frameOptions(HeadersConfigurer.FrameOptionsConfig::disable) // H2 콘솔 프레임 옵션 설정
+			)
+			.addFilterBefore(new JwtAuthenticationFilter(jwtProvider), UsernamePasswordAuthenticationFilter.class);
+		return http.build();
+	}
 
-    @Bean
-    CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration configuration = new CorsConfiguration();
-        configuration.setAllowedOrigins(Arrays.asList(
-                "http://localhost:3000",
-                "https://commitato.site",
-                "https://www.commitato.site"));
-        configuration.setAllowedMethods(Arrays.asList("HEAD", "GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"));
-        configuration.setAllowCredentials(true);
-        configuration.setAllowedHeaders(List.of("*"));
-        configuration.setExposedHeaders(Arrays.asList("Authorization", "Authorization-refresh"));
+	@Bean
+	CorsConfigurationSource corsConfigurationSource() {
+		CorsConfiguration configuration = new CorsConfiguration();
+		configuration.setAllowedOrigins(Arrays.asList(
+			"http://localhost:3000",
+			"http://localhost:8000",
+			"https://commitato.site",
+			"https://www.commitato.site"));
+		configuration.setAllowedMethods(Arrays.asList("HEAD", "GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"));
+		configuration.setAllowCredentials(true);
+		configuration.setAllowedHeaders(List.of("*"));
+		configuration.setExposedHeaders(Arrays.asList("Authorization", "Authorization-refresh"));
 
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-    }
+		UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+		source.registerCorsConfiguration("/**", configuration);
+		return source;
+	}
 }
