+++ src/main/java/com/leets/commitatobe/global/jwt/dto/JwtResponse.java
@@ -1,0 +1,7 @@
+package com.leets.commitatobe.global.jwt.dto;
+
+public record JwtResponse(
+	String grantType,
+	String accessToken
+) {
+}
